Try container renamed to Anon. Anon username of undefined error fixed.
