+++ app/containers/Anon/Anon.scss
@@ -1,0 +1,8 @@
+@import '../../variables';
+
+.Anon {
+  background-color: $background;
+  width:100%;
+  height: 100%;
+  box-sizing: border-box;
+}
