--- app/containers/Try/Try.scss
@@ -1,8 +1,0 @@
-@import '../../variables';
-
-.Try {
-  background-color: $background;
-  width:100%;
-  height: 100%;
-  box-sizing: border-box;
-}
