+++ app/routes.js
@@ -11,7 +11,7 @@ import {
     Projects,
     Recover,
     Signup,
-    Try
+    Anon
   } from './containers';
 export default (
   <Route path="/" component={ App }>
@@ -21,7 +21,7 @@ export default (
     <Route path="recover" component={ Recover } />
     <Route path="signup" component={ Signup }/>
     <Route path="about" component={ About } />
-    <Route path="try/:projectName" component={ Try } />
+    <Route path="anon/:projectName" component={ Anon } />
     <Route path=":owner/:projectName" component={ Project } />
     <Route path=":owner" component={ Projects } />
     <Route path="*" component={ NotFound } />
